test: boardService 이전까지 테스트코드 작성
