+++ src/main/java/org/myteam/server/inquiry/service/InquiryService.java
@@ -3,9 +3,6 @@ package org.myteam.server.inquiry.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.myteam.server.board.domain.Board;
-import org.myteam.server.board.service.BoardCountReadService;
-import org.myteam.server.board.service.BoardRecommendReadService;
 import org.myteam.server.comment.domain.CommentType;
 import org.myteam.server.comment.service.CommentService;
 import org.myteam.server.global.exception.ErrorCode;
@@ -16,10 +13,7 @@ import org.myteam.server.inquiry.repository.InquiryCountRepository;
 import org.myteam.server.inquiry.repository.InquiryRepository;
 import org.myteam.server.member.entity.Member;
 import org.myteam.server.member.repository.MemberJpaRepository;
-import org.myteam.server.member.repository.MemberRepository;
-import org.myteam.server.member.service.MemberReadService;
 import org.myteam.server.member.service.SecurityReadService;
-import org.myteam.server.util.ClientUtils;
 import org.myteam.server.util.slack.service.SlackService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
